backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.0",
    "google-generativeai>=0.8.6",
    "httpx==0.27.0",
    "pydantic==2.9.0",
    "python-dotenv==1.0.1",
    "sqlalchemy==2.0.35",
    "uvicorn==0.30.0",
]
